Refactor: RecommedTask 분리
